Selection sort in C
selection-sort-animationAlgo – Repeatedly finding the minimum element and swapping.

It basically determines the minimum (or maximum) of the list and swaps it with the element at the index where its supposed to be. The process is repeated such that the nth minimum (or maximum) element is swapped with the element at the n-1th index of the list. The below is an implementation of the algorithm in C.